// prisma/generated/zod/inputTypeSchemas/PersonWhereUniqueInputSchema.ts
export const PersonWhereUniqueInputSchema: z.ZodType<Prisma.PersonWhereUniqueInput> = z.object({
  id: z.number()
})
.and(z.object({
  id: z.number().optional(),
  AND: z.union([ z.lazy(() => PersonWhereInputSchema),z.lazy(() => PersonWhereInputSchema).array() ]).optional(),
  OR: z.lazy(() => PersonWhereInputSchema).array().optional(),
  NOT: z.union([ z.lazy(() => PersonWhereInputSchema),z.lazy(() => PersonWhereInputSchema).array() ]).optional(),
  username: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  bio: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  user: z.union([ z.lazy(() => UserNullableRelationFilterSchema),z.lazy(() => UserWhereInputSchema) ]).optional().nullable(),
  priorities: z.lazy(() => PriorityListRelationFilterSchema).optional(),
  projects: z.lazy(() => PersonsToProjectsListRelationFilterSchema).optional(),
  tasks: z.lazy(() => TaskListRelationFilterSchema).optional(),
  skillPaths: z.lazy(() => PersonsToSkillPathsListRelationFilterSchema).optional(),
  skills: z.lazy(() => PersonsToSkillsListRelationFilterSchema).optional(),
  privateEnergies: z.lazy(() => PrivateEnergyListRelationFilterSchema).optional(),
  privateTags: z.lazy(() => PrivateTagListRelationFilterSchema).optional(),
  energies: z.lazy(() => EnergiesToPersonsListRelationFilterSchema).optional(),
  editedProjects: z.lazy(() => PersonsToProjectTemplatesListRelationFilterSchema).optional(),
  editedTasks: z.lazy(() => PersonsToTaskTemplatesListRelationFilterSchema).optional(),
  tagFollows: z.lazy(() => PersonsToTagsListRelationFilterSchema).optional(),
  projectTags: z.lazy(() => ProjectTemplatesToTagsListRelationFilterSchema).optional(),
  taskTags: z.lazy(() => TagsToTaskTemplatesListRelationFilterSchema).optional(),
  skillTags: z.lazy(() => SkillsToTagsListRelationFilterSchema).optional(),
  skillPathTags: z.lazy(() => SkillPathsToTagsListRelationFilterSchema).optional(),
  energyTags: z.lazy(() => EnergiesToTagsListRelationFilterSchema).optional()
}).strict());